move npy-append to dev requirements for conda package mitigation
