+++ requirements/base.txt
@@ -1,2 +1,1 @@
 merlin-core
-npy_append_array
